Add initial implementation of ex-it
Ex-it means "exec and be init".  The idea is to have a tool to exec
another program, wait for it to finish (while reaping children like init
should), and return its exit status.

This is useful in certain scenarios where you want to ensure a
well-behaved init inside a container, but also run only one functional
process inside that container.  See e.g.
https://wiki.gnome.org/Projects/SandboxedApps/Sandbox.

This is untested and I haven't cleaned up the debian/docker bits.  Just
getting ideas out, so far.